# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel V11 Patient and Appointment Management API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/patients": {
            "get": {
                "tags": [
                    "Patients"
                ],
                "summary": "Get all patients",
                "description": "Returns a list of all registered patients",
                "operationId": "c13cef86afd410d31e1d3e75b483dfcb",
                "responses": {
                    "200": {
                        "description": "List of patients successfully obtained"
                    },
                    "204": {
                        "description": "No patients registered"
                    }
                }
            },
            "post": {
                "tags": [
                    "Patients"
                ],
                "summary": "Register a new patient",
                "description": "Registers a patient in the database",
                "operationId": "b2ac8ebf37fd468f3915c3002da3aa51",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "date_born",
                                    "gender",
                                    "address",
                                    "phone"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Juan Pérez"
                                    },
                                    "date_born": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2000-01-01"
                                    },
                                    "gender": {
                                        "type": "string",
                                        "enum": [
                                            "Masculino",
                                            "Femenino"
                                        ],
                                        "example": "Masculino"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Calle Falsa 123"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "12345678"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "juan.perez@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Patient successfully registered"
                    },
                    "400": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/patients/{patientId}": {
            "get": {
                "tags": [
                    "Patients"
                ],
                "summary": "Obtain a patient by ID",
                "description": "Obtains the information of a specific patient by patient ID",
                "operationId": "6ded7d1178c54fde883804809faab439",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Patient ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Patient successfully found"
                    },
                    "404": {
                        "description": "Patient not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Patients"
                ],
                "summary": "Update patient information",
                "description": "Update existing patient data",
                "operationId": "59a13839cf9a82442254de945e70709a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Patient ID to be updated",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "address",
                                    "phone"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Juan Pérez Actualizado"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Avenida Siempre Viva"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "87654321"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "juan.actualizado@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Patient successfully updated"
                    },
                    "400": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter the token returned at login",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}